Refactor VGG16 model loading and preprocessing
This commit refactors the code in the data_preprocess.py file to improve the loading and preprocessing of the VGG16 model. Instead of using the `load_model` method, the code now uses the `from_pretrained` method to load the pretrain model file. Additionally, the encoder is loaded separately from the trained model path. The test data is also preprocessed differently, with the label being transformed using the model's label encoder. These changes aim to enhance the efficiency and accuracy of the VGG16 model.